+++ interface/src/main/java/com/cmpay/lemon/monitor/dto/WorkLoadRspDTO.java
@@ -51,15 +51,15 @@ public class WorkLoadRspDTO extends PageableRspDTO {
     private String riskFeedbackTm;
     private String endMonRemark;
     private String endFeedbackTm;
-    private int totalWorkload;
+    private double totalWorkload;
     //已录入总工作量
-    private int inputWorkload;
+    private double inputWorkload;
     //上月录入
-    private int lastInputWorkload;
+    private double lastInputWorkload;
     //剩余录入工作量
-    private int remainWorkload;
+    private double remainWorkload;
     //本月录入工作量
-    private int monInputWorkload;
+    private double monInputWorkload;
     //	主导部门工作量占比
     private String leadDeptPro;
     //	配合部门工作量占比
@@ -549,43 +549,43 @@ public class WorkLoadRspDTO extends PageableRspDTO {
         this.endFeedbackTm = endFeedbackTm;
     }
 
-    public int getTotalWorkload() {
+    public double getTotalWorkload() {
         return totalWorkload;
     }
 
-    public void setTotalWorkload(int totalWorkload) {
+    public void setTotalWorkload(double totalWorkload) {
         this.totalWorkload = totalWorkload;
     }
 
-    public int getInputWorkload() {
+    public double getInputWorkload() {
         return inputWorkload;
     }
 
-    public void setInputWorkload(int inputWorkload) {
+    public void setInputWorkload(double inputWorkload) {
         this.inputWorkload = inputWorkload;
     }
 
-    public int getLastInputWorkload() {
+    public double getLastInputWorkload() {
         return lastInputWorkload;
     }
 
-    public void setLastInputWorkload(int lastInputWorkload) {
+    public void setLastInputWorkload(double lastInputWorkload) {
         this.lastInputWorkload = lastInputWorkload;
     }
 
-    public int getRemainWorkload() {
+    public double getRemainWorkload() {
         return remainWorkload;
     }
 
-    public void setRemainWorkload(int remainWorkload) {
+    public void setRemainWorkload(double remainWorkload) {
         this.remainWorkload = remainWorkload;
     }
 
-    public int getMonInputWorkload() {
+    public double getMonInputWorkload() {
         return monInputWorkload;
     }
 
-    public void setMonInputWorkload(int monInputWorkload) {
+    public void setMonInputWorkload(double monInputWorkload) {
         this.monInputWorkload = monInputWorkload;
     }
